Move switch and button code to switches.cpp
