+++ switches.cpp
@@ -1,5 +1,5 @@
 /*
- * ukmarsey main file
+ * Switches. Read and test the on-board function switches and button
 
    ukmarsey is a machine and human command-based Robot Low-level I/O platform initially targetting UKMARSBot
    For more information see:
@@ -32,53 +32,60 @@
   SOFTWARE.
 */
 
+#include "switches.h"
 #include "public.h"
-#include "stopwatch.h"
 #include <Arduino.h>
-const int REPORTING_INTERVAL = 10;
-uint32_t report_time_trigger;
-uint8_t PoR_status = 0;
-void setup()
+
+/** @brief  Convert the switch ADC reading into a switch reading.
+ *  @return integer in range 0..16
+ */
+int readFunctionSwitch()
 {
-    PoR_status = MCUSR; // is this erased by bootloader?
-    MCUSR = 0;
-    pinMode(LED_BUILTIN, OUTPUT);
-    Serial.begin(115200);
-    Serial.println(F("Hello from ukmarsey"));
-    setup_systick();
-    sensors_control_setup();
-    setup_encoders();
-    motorSetup();
-    init_stored_parameters();
-    report_time_trigger += REPORTING_INTERVAL;
+    const int adcReading[] = {660, 647, 630, 614, 590, 570, 545, 522, 461,
+                              429, 385, 343, 271, 212, 128, 44, 0};
+
+    if (Switch_ADC_value > 800)
+    {
+        return 16;
+    }
+    for (int i = 0; i < 16; i++)
+    {
+        if (Switch_ADC_value > (adcReading[i] + adcReading[i + 1]) / 2)
+        {
+            return i;
+        }
+    }
+    // TODO: should there be a more informative error value?
+    return 15; // should never get here... but if we do show 15
 }
 
-void loop()
+/** @brief  Test the user pushbutton
+ *  There is no debouncing so take care
+ *  @return boolean
+ */
+bool button_pressed()
 {
-    // fwd_set_speed = 500.0;
-    // rot_set_speed = 0;
-    // handy for simple continuous tests
-    // if (millis() > report_time_trigger)
-    // {
-    //     report_time_trigger += REPORTING_INTERVAL;
-    //     Stopwatch sw;
-    //     Serial.print(millis());
-    //     Serial.print(' ');
-    //     Serial.print(fwd_set_speed);
-    //     Serial.print(' ');
-    //     Serial.print(rot_set_speed);
-    //     Serial.print(' ');
-    //     Serial.print(robot_velocity);
-    //     Serial.print(' ');
-    //     Serial.print(robot_omega);
-    //     Serial.print(' ');
-    //     Serial.print(fwd_volts);
-    //     Serial.print(' ');
-    //     Serial.print(rot_volts); // placeholder for controller voltage
-    //     Serial.print(' ');
-    //     Serial.print(sw.elapsed_time());
-    //     Serial.println();
-    // }
+    return readFunctionSwitch() == 16;
+}
 
-    interpreter();
+void wait_for_button_press()
+{
+    while (not button_pressed())
+    {
+        delay(10);
+    }
+}
+
+void wait_for_button_release()
+{
+    while (button_pressed())
+    {
+        delay(10);
+    }
+}
+
+void wait_for_button_click()
+{
+    wait_for_button_press();
+    wait_for_button_release();
 }
